Exercise_11/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
